purge GPHStatefulRequest and merge it into GPHRequest
